add excel import component and its functionlity like edit or delete and again save excel form and csv form and add csv global service

+++ src/app/order-manage/order-list/order-list.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { CoolorderService } from '../../services/coolorder.service';
-import { response } from 'express';
 import { form, order } from '../../interfaces/form-interface';
 import { CommonModule } from '@angular/common';
 import { Router, RouterLink } from '@angular/router';
@@ -9,14 +8,18 @@ import { LoaderService } from '../../services/loader.service';
 import { HttpContext } from '@angular/common/http';
 import { BYPASS_LOADER } from '../interceptor/loader-context';
 import { ExcelService } from '../../services/excel.service';
-
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { ConfirmBoxComponent } from '../../confirmation-dialog-box/confirm-box/confirm-box.component';
+import * as XLSX from 'xlsx';
+import { CsvService } from '../../services/csv.service';
 
 @Component({
   selector: 'app-order-list',
-  imports: [CommonModule, RouterLink],
+  imports: [CommonModule, RouterLink, ConfirmBoxComponent, MatDialogModule],
   templateUrl: './order-list.component.html',
   styleUrl: './order-list.component.css'
 })
+
 export class OrderListComponent implements OnInit {
 
   public orderData: order[] = [];
@@ -27,46 +30,55 @@ export class OrderListComponent implements OnInit {
   public changeRequest: any
   public successLoader: any = {}
   public approved: any = {}
- 
+  public sortedFiled!: string;
+  public sortType!: string;
 
-
-  constructor(private CoolOrderService: CoolorderService, 
-    public loaderService: LoaderService, 
+  constructor(private CoolOrderService: CoolorderService,
+    public loaderService: LoaderService,
     private router: Router,
-    private ExcelService:ExcelService
-  ) {}
+    private ExcelService: ExcelService,
+    private csvServices:CsvService,
+    private dioLog: MatDialog
+  ) { }
 
   ngOnInit(): void {
-    this.getAllOrder();   
+    this.getAllOrder();
   }
 
-    private getAllOrder() {
-      forkJoin([
-        this.CoolOrderService.getAllOrderData(),
-        this.CoolOrderService.getAllChangeRequest()
-      ]).subscribe(([orders, crs]) => {
-        this.orderData = orders.map((order: order) => {
-          const orderCRs = crs.filter((cr: any) => cr.orderId === order.id.toString());
-          return {
-            ...order,
-            hasCR: orderCRs.length > 0
-          };
-        });
-
-
+  private getAllOrder() {
+    forkJoin([
+      this.CoolOrderService.getAllOrderData(),
+      this.CoolOrderService.getAllChangeRequest()
+    ]).subscribe(([orders, crs]) => {
+      this.orderData = orders.map((order: order) => {
+        const orderCRs = crs.filter((cr: any) => cr.orderId === order.id.toString());
+        return {
+          ...order,
+          hasCR: orderCRs.length > 0
+        };
       });
-    }
+
+
+    });
+  }
 
   public delete(id: any) {
-    this.CoolOrderService.deleteOrder(id).subscribe({
-      next: (response) => {
-        alert("delete successfully")
-        this.getAllOrder();
-      },
-      error: (err) => {
-        alert("order not delete")
+    const dialogBox = this.dioLog.open(ConfirmBoxComponent, { panelClass: 'custom-dialog-container', width: '500px', })
+    dialogBox.afterClosed().subscribe(result => {
+      if (result) {
+
+        this.CoolOrderService.deleteOrder(id).subscribe({
+          next: (response) => {
+            // alert("delete successfully")
+            this.getAllOrder();
+          },
+          error: (err) => {
+            // alert("order not delete")
+          }
+        })
       }
     })
+
   }
 
   public showDialog(orderId: any) {
@@ -222,28 +234,69 @@ export class OrderListComponent implements OnInit {
 
   }
 
-  public short(column:string,  sortType: 'asc' | 'desc'){
-  //  console.log(column, sortType, typeof column, typeof sortType)
-   this.CoolOrderService.getSortItem(column, sortType).subscribe({
-    next: (response)=>{
-      console.log("sort item", response)
-      this.orderData = response
-    }
-   })
+  public short(column: string, sortType: 'asc' | 'desc') {
+    this.sortedFiled = column;
+    this.sortType = sortType
+    this.CoolOrderService.getSortItem(column, sortType).subscribe({
+      next: (response) => {
+        console.log("sort item", response)
+        this.orderData = response
+      }
+    })
   }
 
-  public exportToExcel(){
-   const exportData = this.orderData.map(order => ({
-    OrderID: order.id,
-    Type: order.orderType,
-    From: order.org,
-    To: order.des,
-    RentalDays: order.rentalDays,
-    Status: order.status,
-    leaseStart:order.leaseStart,
-    leaseEnd:order.leaseEnd
-  }));
-  this.ExcelService.exportAsExcelFile(exportData, 'FilteredOrderData')
+  public exportToExcel() {
+    const exportData = this.orderData.map(order => ({
+      org: order.org,
+      des: order.des,
+      pickUpPort: order.pickUpPort,
+      rentalDays: order.rentalDays,
+      leaseStart: order.leaseStart,
+      leaseEnd: order.leaseEnd,
+      productCode: order.productCode,
+      create_at: order.create_at,
+      status: 'draft',
+    }));
+    this.ExcelService.exportAsExcelFile(exportData, 'FilteredOrderData')
   }
+
+  public csv() {
+    console.log("download csv");
+    const exportCvData = this.orderData.map(order => ({
+      Type: order.orderType,
+      From: order.org,
+      To: order.des,
+      RentalDays: order.rentalDays,
+      Status: order.status,
+      leaseStart: order.leaseStart,
+      leaseEnd: order.leaseEnd
+    }));
+
+    this.csvServices.convertToCSV(exportCvData)
+    this.csvServices.exportCSV(exportCvData, 'orderList')
+  }
+
+ 
+
+  // public onFileChange(event: any) {
+  //   console.log("onFile change")
+  //   // console.log(event.target.files[0])
+  //   const file = event.target.files[0];
+  //   const reader = new FileReader();
+
+  //   reader.onload = (e: any) => {
+  //     const binaryData = e.target.result;
+
+  //     const workbook = XLSX.read(binaryData, { type: 'binary' });
+  //     const sheetName = workbook.SheetNames[0];
+  //     const worksheet = workbook.Sheets[sheetName];
+
+  //     this.orderData = XLSX.utils.sheet_to_json(worksheet);
+  //     console.log('Excel JSON Data:', this.orderData);
+  //   };
+
+  //   reader.readAsBinaryString(file);
+  // }
 
 }
+
